move *roll dice* command to new random app
